chore(refactor) moving viewHolderFactory to base package
