+++ app/src/main/java/social/entourage/android/invite/view/InvitationsAdapter.java
@@ -2,7 +2,7 @@ package social.entourage.android.invite.view;
 
 import social.entourage.android.api.model.TimestampedObject;
 import social.entourage.android.base.EntourageBaseAdapter;
-import social.entourage.android.map.tour.information.discussion.ViewHolderFactory;
+import social.entourage.android.base.ViewHolderFactory;
 
 /**
  * Created by mihaiionescu on 08/08/16.
