+++ app/src/main/java/social/entourage/android/map/tour/information/discussion/DiscussionAdapter.java
@@ -5,6 +5,7 @@ import java.util.HashMap;
 
 import social.entourage.android.base.EntourageBaseAdapter;
 import social.entourage.android.api.model.TimestampedObject;
+import social.entourage.android.base.ViewHolderFactory;
 
 /**
  * Created by mihaiionescu on 02/03/16.
