+++ app/src/main/java/social/entourage/android/newsfeed/NewsfeedAdapter.java
@@ -6,7 +6,7 @@ import social.entourage.android.api.model.TimestampedObject;
 import social.entourage.android.map.MapViewHolder;
 import social.entourage.android.map.entourage.EntourageViewHolder;
 import social.entourage.android.map.tour.TourViewHolder;
-import social.entourage.android.map.tour.information.discussion.ViewHolderFactory;
+import social.entourage.android.base.ViewHolderFactory;
 
 /**
  * Created by mihaiionescu on 05/05/16.
